float value

+++ backend/models/product_model.py
@@ -5,9 +5,9 @@ from django.db import models
 class product(models.Model):
     branch = models.CharField(max_length=200)
     product_type = models.CharField(max_length=100)
-    calories  = models.IntegerField(default=0)
-    saturated_fats_percentage = models.IntegerField(default=0)
-    sugar_percentage = models.IntegerField(default=0)
+    calories  = models.FloatField(default=0)
+    saturated_fats_percentage = models.FloatField(default=0)
+    sugar_percentage = models.FloatField(default=0)
 
 
     def __str__(self):
